chore: refactoring
- 인기강의, 신규강의 중복부분 컴포넌트로 분리
- 앱 전반적으로 쓰이는 타입들은 @typing/common 파일로 분리
- tsconfig 수정
