+++ src/components/Main/Header/types.ts
@@ -1,0 +1,3 @@
+export type SearchBoxProps = { placeholder: string };
+export type AlarmProps = { onPress: () => void; hasAlarm: boolean };
+export type FilterProps = { onPress: () => void };
