+++ src/components/Main/index.tsx
@@ -1,4 +1,3 @@
-export { default as Header } from './Header';
-export { default as NewLectureList } from './NewLectureList';
-export { default as PopularLectureList } from './PopularLectureList';
-export { default as SearchBox } from './SearchBox';
+export { default as Header } from './Header/Header';
+export { default as NewLectureList } from './Lecture/NewLectureList';
+export { default as PopularLectureList } from './Lecture/PopularLectureList';
